Con la parte de archivos

+++ Calendario2/Controllers/ArchivosController.cs
@@ -83,15 +83,12 @@ namespace INTRANETVOID.Controllers
 
                     string currentUserId = User.Identity.GetUserId();
                     var IdUser = currentUserId;
-
-
-
                     string _FileName = Path.GetFileName(upload.FileName);
                         tema = (from t in db.Temas
                                        where t.Id.ToString() == archivo.IdTema.ToString()
                                        select t.Descripcion).First().ToString();
 
-                        string strPath = "~/Views/Shared/Upload/" + tema.ToString();
+                        string strPath = "~/Views/Shared/Upload/" + archivo.IdTema.ToString();
                         _path = Path.Combine(Server.MapPath(strPath), _FileName);
                         // se crea directorio si no hay archivos para ese tema, es el primer archivo que se carga
                         int cuantos = (from s in db.Archivos
@@ -112,9 +109,6 @@ namespace INTRANETVOID.Controllers
                         str_desc = "Crea archivo - " + archivo.Nombre.ToString() + " - en carpeta -" + tema.ToString() + " -";
                         //var proc = db.pa_crear_log(1, str_desc, 1, 1, archivo.IdCliente, 1, 0, System.DateTime.Now);
 
-
-
-
                         ViewBag.Message = "Archivo cargado exitosamente!!" + str_desc.ToString();
                     //}
                     //catch
@@ -128,8 +122,12 @@ namespace INTRANETVOID.Controllers
             //ViewBag.IdCliente = new SelectList(db.Clientes.OrderBy(s => s.Nombres), "Id", "Nombres", archivo.IdCliente);
             ViewBag.IdTema = new SelectList(db.Temas.OrderBy(s => s.Descripcion), "Id", "Nombre", archivo.IdTema);
             ViewBag.Message = ViewBag.Message + "-->> Y se creo correctamente el registro en la base de datos!!";
-           //return  RedirectToAction("Index", new { tema });
-            return View(archivo);
+            //return  RedirectToAction("Index", new { tema });
+            TempData["Msg"] = ViewBag.Message;
+
+            return RedirectToAction("Index", new { Id = archivo.IdTema, archivo.Nombre });
+
+            //return View(archivo);
 
         }
 
@@ -177,6 +175,7 @@ namespace INTRANETVOID.Controllers
             }
             Archivo archivo = db.Archivos.Find(id);
             ViewBag.Nombre = Nombre;
+            ViewBag.IdTema = archivo.IdTema;
             if (archivo == null)
             {
                 return HttpNotFound();
@@ -190,24 +189,25 @@ namespace INTRANETVOID.Controllers
         public ActionResult DeleteConfirmed(int id,string Nombre)
         {
             Archivo archivo = db.Archivos.Find(id);
-            string fullPath = Request.MapPath("~/Views/Shared/Upload/" + Nombre + "/" + archivo.Nombre.ToString());
+            string fullPath = Request.MapPath("~/Views/Shared/Upload/" + archivo.IdTema.ToString() + "/" + archivo.Nombre.ToString());
             var str_desc = "";
             var cuantos = 0;
+            var IdTema = "";
             if (System.IO.File.Exists(fullPath))
             {
                 System.IO.File.Delete(fullPath);
-                str_desc = "Elimina archivo - " + archivo.Nombre.ToString() + " - de carpeta -" + Nombre.ToString() + " -";
+                str_desc = "Elimina archivo - " + archivo.Nombre.ToString() + " - de carpeta -" + archivo.IdTema.ToString() + " -";
 
 
-                var IdTema = archivo.IdTema;
+                IdTema = archivo.IdTema.ToString();
                 // ** Si es el ultimo archivo debe borrar la carpeta
                 cuantos = (from s in db.Archivos
-                               where s.IdTema == IdTema
+                               where s.IdTema.ToString() == IdTema.ToString()
                                select s).Count();
             }
             else
             {
-                str_desc = "No se encontró el archivo en el sistema.  Se Elimina archivo de la base de datos - " + archivo.Nombre.ToString() + " - de carpeta -" + Nombre.ToString() + " -";
+                str_desc = "No se encontró el archivo en el sistema.  Se Elimina archivo de la base de datos - " + archivo.Nombre.ToString() + " - de carpeta -" + archivo.IdTema.ToString() + " -";
             }
             // logs
             //var proc = db.pa_crear_log(3, str_desc, 1, 1, archivo.IdCliente, 3, 0, System.DateTime.Now);
@@ -215,14 +215,35 @@ namespace INTRANETVOID.Controllers
             db.Archivos.Remove(archivo);
             db.SaveChanges();
             ViewBag.Nombre = Nombre;
+            ViewBag.IdTema = archivo.IdTema;
+            var strPath2 = "";
             // ** Si es el ultimo archivo debe borrar la carpeta
             if (cuantos == 1)
             {
 
-                string strPath = "~/Views/Shared/Upload/" + Nombre;
-                var _path = Path.Combine(Server.MapPath(strPath),"");
+                string strPath = "~/Views/Shared/Upload/" + archivo.IdTema.ToString();
+                strPath2 = Path.Combine(Server.MapPath(strPath), "");
+                //strPath2 = Server.MapPath(strPath);
 
-                Directory.Delete(_path);
+                if (Directory.Exists(strPath2))
+                {
+                    var cantf = Directory.EnumerateFiles(strPath2).Count();
+                    if (cantf == 0)
+                    //    var _path = Path.Combine(Server.MapPath(strPath),"");
+                    {
+                        var _path = Path.Combine(Server.MapPath(strPath));
+                        Directory.Delete(_path);
+                        str_desc = str_desc + "Se borro el directorio SATISFACTORIAMENTE! ";
+                    }
+                    else
+                    {
+                        str_desc = str_desc + "No se pudo borrar el directorio - " + archivo.IdTema.ToString() + " - por que no esta vacio ";
+                    }
+                }
+                else
+                {
+                    str_desc = str_desc + " No se encontró el directorio - " + strPath2 + " - a borrar";
+                }
             }
             TempData["Msg"] = str_desc;
             return RedirectToAction("Index",new {Id=archivo.IdTema, Nombre });
